+++ signer/src/storage/mod.rs
@@ -49,15 +49,15 @@ pub trait DbRead {
     fn get_pending_deposit_requests(
         &self,
         chain_tip: &model::BitcoinBlockHash,
-        context_window: i32,
+        context_window: u16,
     ) -> impl Future<Output = Result<Vec<model::DepositRequest>, Self::Error>> + Send;
 
     /// Get pending deposit requests that have been accepted by at least `threshold` signers and has no responses
     fn get_pending_accepted_deposit_requests(
         &self,
         chain_tip: &model::BitcoinBlockHash,
-        context_window: i32,
-        threshold: i64,
+        context_window: u16,
+        threshold: u16,
     ) -> impl Future<Output = Result<Vec<model::DepositRequest>, Self::Error>> + Send;
 
     /// Get the deposit requests that the signer has accepted to sign
@@ -70,13 +70,13 @@ pub trait DbRead {
     fn get_deposit_signers(
         &self,
         txid: &model::BitcoinTxId,
-        output_index: i32,
+        output_index: u32,
     ) -> impl Future<Output = Result<Vec<model::DepositSigner>, Self::Error>> + Send;
 
     /// Get signer decisions for a withdraw request
     fn get_withdraw_signers(
         &self,
-        request_id: i32,
+        request_id: u64,
         block_hash: &model::StacksBlockHash,
     ) -> impl Future<Output = Result<Vec<model::WithdrawSigner>, Self::Error>> + Send;
 
@@ -84,15 +84,15 @@ pub trait DbRead {
     fn get_pending_withdraw_requests(
         &self,
         chain_tip: &model::BitcoinBlockHash,
-        context_window: i32,
+        context_window: u16,
     ) -> impl Future<Output = Result<Vec<model::WithdrawRequest>, Self::Error>> + Send;
 
     /// Get pending withdraw requests that have been accepted by at least `threshold` signers and has no responses
     fn get_pending_accepted_withdraw_requests(
         &self,
         chain_tip: &model::BitcoinBlockHash,
-        context_window: i32,
-        threshold: i64,
+        context_window: u16,
+        threshold: u16,
     ) -> impl Future<Output = Result<Vec<model::WithdrawRequest>, Self::Error>> + Send;
 
     /// Get bitcoin blocks that include a particular transaction
