+++ signer/src/storage/postgres.rs
@@ -272,7 +272,7 @@ impl super::DbRead for PgStore {
     async fn get_pending_deposit_requests(
         &self,
         chain_tip: &model::BitcoinBlockHash,
-        context_window: i32,
+        context_window: u16,
     ) -> Result<Vec<model::DepositRequest>, Self::Error> {
         sqlx::query_as::<_, model::DepositRequest>(
             r#"
@@ -312,7 +312,7 @@ impl super::DbRead for PgStore {
             "#,
         )
         .bind(chain_tip)
-        .bind(context_window)
+        .bind(context_window as i32)
         .fetch_all(&self.0)
         .await
         .map_err(Error::SqlxQuery)
@@ -321,8 +321,8 @@ impl super::DbRead for PgStore {
     async fn get_pending_accepted_deposit_requests(
         &self,
         chain_tip: &model::BitcoinBlockHash,
-        context_window: i32,
-        threshold: i64,
+        context_window: u16,
+        threshold: u16,
     ) -> Result<Vec<model::DepositRequest>, Self::Error> {
         sqlx::query_as::<_, model::DepositRequest>(
             r#"
@@ -370,8 +370,8 @@ impl super::DbRead for PgStore {
             "#,
         )
         .bind(chain_tip)
-        .bind(context_window)
-        .bind(threshold)
+        .bind(context_window as i32)
+        .bind(threshold as i32)
         .fetch_all(&self.0)
         .await
         .map_err(Error::SqlxQuery)
@@ -409,7 +409,7 @@ impl super::DbRead for PgStore {
     async fn get_deposit_signers(
         &self,
         txid: &model::BitcoinTxId,
-        output_index: i32,
+        output_index: u32,
     ) -> Result<Vec<model::DepositSigner>, Self::Error> {
         sqlx::query_as::<_, model::DepositSigner>(
             "SELECT
@@ -422,7 +422,7 @@ impl super::DbRead for PgStore {
             WHERE txid = $1 AND output_index = $2",
         )
         .bind(txid)
-        .bind(output_index)
+        .bind(output_index as i32)
         .fetch_all(&self.0)
         .await
         .map_err(Error::SqlxQuery)
@@ -430,7 +430,7 @@ impl super::DbRead for PgStore {
 
     async fn get_withdraw_signers(
         &self,
-        request_id: i32,
+        request_id: u64,
         block_hash: &model::StacksBlockHash,
     ) -> Result<Vec<model::WithdrawSigner>, Self::Error> {
         sqlx::query_as::<_, model::WithdrawSigner>(
@@ -443,7 +443,7 @@ impl super::DbRead for PgStore {
             FROM sbtc_signer.withdraw_signers
             WHERE request_id = $1 AND block_hash = $2",
         )
-        .bind(request_id)
+        .bind(request_id as i64)
         .bind(block_hash)
         .fetch_all(&self.0)
         .await
@@ -453,7 +453,7 @@ impl super::DbRead for PgStore {
     async fn get_pending_withdraw_requests(
         &self,
         chain_tip: &model::BitcoinBlockHash,
-        context_window: i32,
+        context_window: u16,
     ) -> Result<Vec<model::WithdrawRequest>, Self::Error> {
         let Some(stacks_chain_tip) = self.get_stacks_chain_tip(chain_tip).await? else {
             return Ok(Vec::new());
@@ -522,7 +522,7 @@ impl super::DbRead for PgStore {
         )
         .bind(chain_tip)
         .bind(stacks_chain_tip)
-        .bind(context_window)
+        .bind(context_window as i32)
         .fetch_all(&self.0)
         .await
         .map_err(Error::SqlxQuery)
@@ -531,8 +531,8 @@ impl super::DbRead for PgStore {
     async fn get_pending_accepted_withdraw_requests(
         &self,
         chain_tip: &model::BitcoinBlockHash,
-        context_window: i32,
-        threshold: i64,
+        context_window: u16,
+        threshold: u16,
     ) -> Result<Vec<model::WithdrawRequest>, Self::Error> {
         let Some(stacks_chain_tip) = self.get_stacks_chain_tip(chain_tip).await? else {
             return Ok(Vec::new());
@@ -608,8 +608,8 @@ impl super::DbRead for PgStore {
         )
         .bind(chain_tip)
         .bind(stacks_chain_tip)
-        .bind(context_window)
-        .bind(threshold)
+        .bind(context_window as i32)
+        .bind(threshold as i64)
         .fetch_all(&self.0)
         .await
         .map_err(Error::SqlxQuery)
@@ -752,7 +752,7 @@ impl super::DbWrite for PgStore {
             ON CONFLICT DO NOTHING",
         )
         .bind(&block.block_hash)
-        .bind(block.block_height)
+        .bind(block.block_height as i64)
         .bind(&block.parent_hash)
         .bind(&block.confirms)
         .execute(&self.0)
@@ -773,7 +773,7 @@ impl super::DbWrite for PgStore {
             ON CONFLICT DO NOTHING",
         )
         .bind(&block.block_hash)
-        .bind(block.block_height)
+        .bind(block.block_height as i64)
         .bind(&block.parent_hash)
         .execute(&self.0)
         .await
@@ -801,12 +801,12 @@ impl super::DbWrite for PgStore {
             ON CONFLICT DO NOTHING",
         )
         .bind(&deposit_request.txid)
-        .bind(deposit_request.output_index)
+        .bind(deposit_request.output_index as i32)
         .bind(&deposit_request.spend_script)
         .bind(&deposit_request.reclaim_script)
         .bind(&deposit_request.recipient)
-        .bind(deposit_request.amount)
-        .bind(deposit_request.max_fee)
+        .bind(deposit_request.amount as i64)
+        .bind(deposit_request.max_fee as i64)
         .bind(&deposit_request.sender_addresses)
         .execute(&self.0)
         .await
@@ -834,12 +834,12 @@ impl super::DbWrite for PgStore {
 
         for req in deposit_requests {
             txid.push(req.txid);
-            output_index.push(req.output_index);
+            output_index.push(req.output_index as i32);
             spend_script.push(req.spend_script);
             reclaim_script.push(req.reclaim_script);
             recipient.push(req.recipient);
-            amount.push(req.amount);
-            max_fee.push(req.max_fee);
+            amount.push(req.amount as i64);
+            max_fee.push(req.max_fee as i64);
             // We need to join the addresses like this (and later split
             // them), because handling of multidimensional arrays in
             // postgres is tough. The naive approach of doing
@@ -917,11 +917,11 @@ impl super::DbWrite for PgStore {
             VALUES ($1, $2, $3, $4, $5, $6)
             ON CONFLICT DO NOTHING",
         )
-        .bind(withdraw_request.request_id)
+        .bind(withdraw_request.request_id as i64)
         .bind(&withdraw_request.block_hash)
         .bind(&withdraw_request.recipient)
-        .bind(withdraw_request.amount)
-        .bind(withdraw_request.max_fee)
+        .bind(withdraw_request.amount as i64)
+        .bind(withdraw_request.max_fee as i64)
         .bind(&withdraw_request.sender_address)
         .execute(&self.0)
         .await
@@ -946,7 +946,7 @@ impl super::DbWrite for PgStore {
             ON CONFLICT DO NOTHING",
         )
         .bind(&decision.txid)
-        .bind(decision.output_index)
+        .bind(decision.output_index as i32)
         .bind(decision.signer_pub_key)
         .bind(decision.is_accepted)
         .execute(&self.0)
@@ -970,7 +970,7 @@ impl super::DbWrite for PgStore {
             VALUES ($1, $2, $3, $4)
             ON CONFLICT DO NOTHING",
         )
-        .bind(decision.request_id)
+        .bind(decision.request_id as i64)
         .bind(&decision.block_hash)
         .bind(decision.signer_pub_key)
         .bind(decision.is_accepted)
@@ -1123,7 +1123,7 @@ impl super::DbWrite for PgStore {
         for block in blocks {
             block_ids.push(block.block_hash);
             parent_block_ids.push(block.parent_hash);
-            chain_lengths.push(block.block_height);
+            chain_lengths.push(block.block_height as i64);
         }
 
         sqlx::query(
